feat: add sorting and searching
